+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## Unreleased
+## 1.7.0
 
 * Add `classes` prop types
 * Add `component` prop type to `Button` component
 * Add `legacy` and `future` themes
+* Change default theme to `legacy`
 
 ## 1.6.4 (2018-12-12)